Get delete working again
This uncovered a bug in my streak code where I get stuck in an infinite
loop.  My bad.  But this made the async delete task hard to debug.
Also, while I could remove the row from the database, it did not seem to
refresh the listview notifyDataSetChanged() did nothing, had to restart
the loader.